% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_bmi_column}
\alias{add_bmi_column}
\title{Add BMI column}
\usage{
add_bmi_column(df, height_cm, weight_kg)
}
\arguments{
\item{df}{Input data.frame with 'height_cm' and 'weight_kg' columns}
}
\value{
data.frame with added 'bmi' column
}
\description{
Add a BMI column to the data.frame
}
