% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{Calculate BMI}
\usage{
calculate_bmi(height_cm, weight_kg)
}
\arguments{
\item{height_cm}{Height in centimeters}

\item{weight_kg}{Weight in kilograms}
}
\value{
float: Calculated BMI
}
\description{
Calculate BMI from height (cm) and weight (kg)
}
