% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{calculate_disease_prevalence}
\alias{calculate_disease_prevalence}
\title{Calculate disease prevalence rates per month}
\usage{
calculate_disease_prevalence(df)
}
\arguments{
\item{df}{cleaned health data data.frame}
}
\value{
data.frame with prevalence rates per month and diagnosis
}
\description{
Calculate disease prevalence rates per month from
cleaned health data
}
