% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_health_data}
\alias{clean_health_data}
\title{Clean health data}
\usage{
clean_health_data(df)
}
\arguments{
\item{df}{raw health data data.frame}
}
\value{
df_clean cleaned health data data.frame
}
\description{
Clean health data by handling missing values and standardizing columns
}
