% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{format_month_section}
\alias{format_month_section}
\title{Format the markdown section for a single month}
\usage{
format_month_section(month, month_df)
}
\arguments{
\item{month}{Month string}

\item{month_df}{data.frame filtered for the month}
}
\value{
Markdown string for the month section
}
\description{
Format the markdown section for a single month
}
