% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_markdown_report}
\alias{generate_markdown_report}
\title{Generate a markdown report of disease prevalence rates per month.}
\usage{
generate_markdown_report(prevalence_df, output_path)
}
\arguments{
\item{prevalence_df}{data.frame with prevalence rates.}

\item{output_path}{Path to output markdown file.}
}
\description{
Generate a markdown report of disease prevalence rates per month.
}
