% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_config}
\alias{load_config}
\title{Load pipeline configuration files}
\usage{
load_config(config_path)
}
\arguments{
\item{config_path}{Path to the YAML config file}
}
\value{
dict Configuration dictionary
}
\description{
Load pipeline configuration from a YAML file
}
