% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_health_data}
\alias{read_health_data}
\title{Read health data CSV into a data.frame}
\usage{
read_health_data(filepath)
}
\arguments{
\item{filepath}{Path to the CSV file}
}
\value{
Data frame containing the health data
}
\description{
Read a CSV file containing health data into a data.frame
}
