% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{write_dataframe}
\alias{write_dataframe}
\title{Write data.frame to CSV}
\usage{
write_dataframe(df, filepath)
}
\arguments{
\item{df}{data.frame to write}

\item{filepath}{Path to the output CSV file}
}
\description{
Write a data.frame to a CSV file
}
